Add files via upload
Requirements:
Create a bag data structure:

The user inserts key-value pair.

Each key is related to a value.

To insert, the user needs to insert both key and value. If key is already there, the value will be replaced.

To retrieve, the user inserts key and receives the value.

To remove, the user inserts key and the key value pair will be removed.

Modify your bag so the insert does not happen at the end but in a sorted order.

Modify your other two functions to make them faster.

So far we assumed both key and values are integers. What happens if key is integer and the value is a student object.

Modify your program to work with the new settings.

* I assume the students are familiar with the topic of binary search.